fix(@quasar/app): Ensure fixed deps in electron build package.json
